# website/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mitabl eComm API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/restaurant/menu/{resturantId}": {
            "get": {
                "tags": [
                    "kitchen"
                ],
                "summary": "Restaurant Menu",
                "description": "Returns Food data",
                "operationId": "restaurant Menu",
                "parameters": [
                    {
                        "name": "resturantId",
                        "in": "path",
                        "description": "Resturant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "All details fetched Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "All details fetched Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v1/food/status/{id}": {
            "get": {
                "tags": [
                    "kitchen"
                ],
                "summary": "change food item status",
                "description": "Returns Food data",
                "operationId": "change food item status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Food id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "All details fetched Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "All details fetched Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}